add documentation and tests

+++ config/initializers/reset_presences.rb
@@ -1,2 +1,2 @@
 # Reset all presences to start with empty chat rooms
-Presence.delete_all
+Presence.delete_all if Presence.table_exists?
